Fixed dataout writebus socket to being bytes instead of int

+++ TCPIP/src/Memory/BusDefinitions.cs
@@ -13,7 +13,7 @@ namespace TCPIP
         {
             long frame_number { get; set; }
             int socket { get; set; }
-            int data { get; set; }
+            byte data { get; set; }
         }
     }
